% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{methods}
\name{Chrom}
\alias{Chrom}
\alias{Chrom,vcfLink-method}
\title{Extract SNP ids and positions}
\usage{
Chrom(object, ...)

\S4method{Chrom}{vcfLink}(object)
}
\arguments{
\item{object}{an object of class "vcfLink"}
}
\description{
Given a "vcfLink" object, returns a data frame with the chromosome (contig), position, and ID of each SNP.
}
