% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{Copy}
\alias{Copy}
\alias{Copy,vcfLink-method}
\title{Copy vcfLink object}
\usage{
Copy(object, ...)

\S4method{Copy}{vcfLink}(object)
}
\arguments{
\item{object}{an object of class "vcfLink"}
}
\description{
Copies a vcfLink object, and creates a copy of the temporary (working) VCF file for it to link to
}

