% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{Filter}
\alias{Filter}
\alias{Filter,vcfLink-method}
\alias{filterOptions}
\title{Filter VCF files}
\usage{
Filter(object, ...)

\S4method{Filter}{vcfLink}(object, arguments = filterOptions(),
  indels = c("keep", "remove", "only"), verbose = TRUE)

filterOptions(min.meanDP = NULL, max.meanDP = NULL, minDP = NULL,
  maxDP = NULL, min.alleles = NULL, max.alleles = NULL, maf = NULL,
  max.maf = NULL, minQ = NULL, max.missing = NULL, hwe = NULL,
  thin = NULL)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{arguments}{optional arguments for vcftools as returned by \code{filterOptions}, see vcftools manual.}

\item{indels}{what to do with indels; "keep" leaves them in, "remove" removes them, and "only" discards everything but indels.}

\item{verbose}{print stdout from vcftools?}
}
\description{
Given a "vcfLink" object, removes loci/samples corresponding to a set of filtering options.
}
