% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\docType{methods}
\name{Fstats}
\alias{Fstats}
\alias{Fstats,vcfLink-method}
\title{Fst calculation from VCF file}
\usage{
Fstats(object, ...)

\S4method{Fstats}{vcfLink}(object, by, mean_Fst = TRUE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{by}{a vector containing population assignments for each sample}

\item{mean_Fst}{return an estimate of Fst averaged over loci, or separate Fst values for each locus?}
}
\description{
Given a "vcfLink" object and a list of population IDs, returns a data frame containing pairwise Fst
}
