% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{Geno}
\alias{Geno}
\alias{Geno,vcfLink-method}
\title{Convert VCF to genotype matrix}
\usage{
Geno(object, ...)

\S4method{Geno}{vcfLink}(object, output.file, verbose = TRUE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{output.file}{the filename to save the genotype matrix to}

\item{verbose}{print stdout from vcftools?}
}
\description{
Given a "vcfLink" object, converts the VCF file to genotype matrix format. Doing so involves dropping indels and multiallelic SNPs, so this fuction returns a new vcfLink object that has been appropriately subset.
}
