% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{Lfmm}
\alias{Lfmm}
\alias{Lfmm,vcfLink-method}
\title{Convert VCF to LFMM format}
\usage{
Lfmm(object, ...)

\S4method{Lfmm}{vcfLink}(object, output.file, verbose = TRUE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{output.file}{the filename to save the LFMM file to}

\item{verbose}{print stdout from vcftools?}
}
\description{
Given a "vcfLink" object, converts the VCF file to the format used by LFMM in package LEA. Involves intermediate creation of a genotype matrix, so only biallelic SNPs are retained.
}
