% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{LfmmEnv}
\alias{LfmmEnv}
\alias{LfmmEnv,vcfLink-method}
\title{Create environmenal variable file for LFMM}
\usage{
LfmmEnv(object, ...)

\S4method{LfmmEnv}{vcfLink}(object, varnames, output.file = NULL)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{varnames}{a character vector of the variables in the metadata to write to the .env file}

\item{output.file}{the filename to save the ENV file to}
}
\description{
Given a "vcfLink" object, writes given variables in the metadata to a .env file used by LFMM in package LEA.
}

