% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{methods}
\name{Linkage}
\alias{Linkage}
\alias{Linkage,vcfLink-method}
\alias{linkageOptions}
\title{Calculate linkage disequillibrium between loci}
\usage{
Linkage(object, ...)

\S4method{Linkage}{vcfLink}(object, type = c("geno-r2", "geno-chisq",
  "interchrom-geno-r2"), arguments = linkageOptions(), verbose = TRUE)

linkageOptions(min.r2 = NULL, ld.window = NULL, ld.window.bp = NULL,
  ld.window.min = NULL, ld.window.bp.min = NULL)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{type}{the type of comparison to make. See vcftools manual.}

\item{arguments}{optional arguments for vcftools as returned by \code{linkageOptions}, see vcftools manual.}

\item{verbose}{print stdout from vcftools?}
}
\description{
Given a "vcfLink" object, returns a data frame with pairwise p-values between loci, under the null hypothesis that the loci are not in linkage disequillibrium.
}

