% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{LoadMeta}
\alias{LoadMeta}
\alias{LoadMeta,vcfLink-method}
\title{Attach metadata to vcfLink object}
\usage{
LoadMeta(object, ...)

\S4method{LoadMeta}{vcfLink}(object, meta)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{meta}{a "data.frame" containing sample metadata. Columns correspond to variables, rows correspond to samples.}
}
\description{
Attaches a dataframe containing sample metadata to a vcfLink object.
}

