% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Query}
\alias{Query}
\title{Query attributes of a VCF file}
\usage{
Query(object, ...)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{type}{the type of information to return; see the vcftools manual for details}

\item{verbose}{boolean; if true, return stdout from vcftools}
}
\description{
Use vcftools to extract information from a VCF file.
}

