% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\docType{methods}
\name{Relatedness}
\alias{Relatedness}
\alias{Relatedness,vcfLink-method}
\title{Relatedness calculation from VCF file}
\usage{
Relatedness(object, ...)

\S4method{Relatedness}{vcfLink}(object, type = c("yang", "manichaikul"),
  verbose = TRUE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{type}{the type of relatedness coefficient to return; see vcftools manual}

\item{verbose}{print stdout from vcftools?}
}
\description{
Given a "vcfLink" object, returns a data frame containing pairwise relatedness coefficients between samples.
}

