% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{Save}
\alias{Save}
\alias{Save,vcfLink-method}
\title{Save vcfLink object}
\usage{
Save(object, ...)

\S4method{Save}{vcfLink}(object, filename)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{filename}{character; the path to which to save the VCF file to}
}
\description{
Saves the temporary VCF linked to by a vcfLink object to a given filename. Also, saves any sample metadata which is associated with the VCF file.
}

