% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{Subset}
\alias{Subset}
\alias{Subset,vcfLink-method}
\title{Subset VCF files}
\usage{
Subset(object, ...)

\S4method{Subset}{vcfLink}(object, samples = NULL, sites = NULL,
  verbose = TRUE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{samples}{a character vector containing either positions or names of samples to keep}

\item{verbose}{print stdout from vcftools?}

\item{site}{a character vector containing either positions or names of sites to keep}
}
\description{
Given a "vcfLink" object, subsets loci/samples to a given vector of sample or site IDs.
}

