% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_d.R
\name{TajimaD}
\alias{TajimaD}
\title{Tajima's D}
\usage{
TajimaD(object, maf = 0, nboot = 0, nsim = 1000, ploidy = 2,
  use_vcftools_D = FALSE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{maf}{the minimum allele frequency allowed in the null model simulations, as a proportion; SNPs with frequency < MAF are excluded from the simulations. The data is also thinned using this criterion prior to calculating D.}

\item{nboot}{the number of bootstrap replicates used to generate a confidence interval around D; if 0, no bootstrap is performed}

\item{nsim}{the number of simulations from the null model, used to correct bias due to the MAF and to test the significance of D; if 0, no correction is performed}

\item{ploidy}{the ploidy of the organism}

\item{use_vcftools_D}{if TRUE, use the method for calculating D implemented in vcftools, which is biased when there is missing data. If FALSE (default), use an unbiased (and faster) method}
}
\description{
Calculate a genome-wide estimate of Tajima's D, and optionally use simulation from the neutral model to correct for bias due to a minor-allele-frequency filter and perform a significance test.
}
