% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{UpdateLink}
\alias{UpdateLink}
\alias{UpdateLink,vcfLink-method}
\title{Update link to temporary VCF}
\usage{
UpdateLink(object, newlink, ...)

\S4method{UpdateLink}{vcfLink}(object, newlink, clean = TRUE)
}
\arguments{
\item{object}{an object of class "vcfLink"}

\item{newlink}{character; new path to update object with}

\item{clean}{boolean; clean temporary directory? If true, the old file that the object is linked to will be deleted}
}
\description{
Links a vcfLink to a new temporary path
}

