% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcflink.R
\docType{class}
\name{vcfLink-class}
\alias{vcfLink-class}
\title{An S4 class containing a link to a VCF file which vcftools interacts with.}
\description{
An S4 class containing a link to a VCF file which vcftools interacts with.
}
\section{Slots}{

\describe{
\item{\code{vcf_path}}{The path to the original VCF file}

\item{\code{vcf_scratch}}{The path to the temporary (working) VCF file}

\item{\code{meta_path}}{The path to a csv file containing metadata describing the samples}

\item{\code{sample_id}}{A character vector containing the names of samples}

\item{\code{meta}}{A data frame containing metadata describing the samples}

\item{\code{site_id}}{A character vector containing the names of SNPs}
}}

