% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcflink.R
\name{vcfLink}
\alias{vcfLink}
\title{Create a vcfLink object}
\usage{
vcfLink(vcf_path, meta = NULL, overwriteID = FALSE)
}
\arguments{
\item{vcf_path}{Character, the path to a VCF file}

\item{meta}{Character, the path to a csv containing metadata associated with each sample}

\item{overwriteID}{Boolean, overwrite SNP ids with numbers? Only set to TRUE if the VCF file does not already contain SNP ids}
}
\description{
Creates an object of class vcfLink, which acts as an interface to a temporary VCF file. Also loads metadata (if present) associated with the samples.
}

